Messy Commit, Final Publish of code
I accidentally made a commit, then got myself tangled trying to undo the
commit then trying to reset, so i just reset the whole HEAD to where I
wanted and moved on from there. Also, these are the final versions of
the files. Maybe I'll build an auto-mailer or something.